pyproject.toml
[tool.poetry]
name = "textable-backend"
version = "0.1.0"
description = ""
authors = ["Bharath Sriraam R R <albharath.1305@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-cloud-translate = "2.0.1"
fastapi = "^0.110.1"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
python-dotenv = "^1.0.1"
langchain = "^0.1.16"
langchain-openai = "^0.1.3"
instructor = "^1.2.0"
openai = "^1.17.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
